diploma-backend/megano/pyproject.toml
[project]
name = "megano"
version = "0.1.0"
description = ""
authors = [
    {name = "user",email = "ilya.melnichuk.2007@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==4.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "diploma-frontend @ file:///home/user/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
